Split up events tests into individual test classes
